feat: add Kubevirt deployment
This adds the automated deployment of Kubevirt when
kubevirt_enable is set to true.
The Kubevirt version can be defined via the kubevirt_version
variable.
As the Kubevirt Kubernetes YAMLs are downloaded from the release
page directly, any version of Kubevirt that has the kubevirt-cr.yaml
and kubevirt-operator.yaml file should work.